little fix

+++ modules/core/src/intrinsic.cpp
@@ -32,7 +32,7 @@ uint8_t log2(uint32_t value) noexcept {
 #if IMAGESTEGO_CLANG || IMAGESTEGO_GCC
     return value ? 31 - __builtin_clz(value) : 0;
 #elif IMAGESTEGO_MSVC || IMAGESTEGO_ICC
-    unsigned long result = 0;
+    uint32_t result = 0;
     _BitScanReverse(&result, value);
     return result;
 #else
